Moving the plotstyle function into a separate module, which is now used by all interactive plots
